replace all courses int to long type and remove CourseId in CourseReq

+++ src/StudyThink.Service/DTOs/Courses/CourseComment/CourseCommentCreationDto.cs
@@ -3,7 +3,7 @@
 public class CourseCommentCreationDto
 {
     public string Comment { get; set; }
-    public int StudentId { get; set; }
-    public int CourseId { get; set; }
-    public int AdminId { get; set; }
+    public long StudentId { get; set; }
+    public long CourseId { get; set; }
+    public long AdminId { get; set; }
 }
